project is moved as a part of high level project
