--- .idea/compiler.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <annotationProcessing>
-      <profile name="Maven default annotation processors profile" enabled="true">
-        <sourceOutputDir name="target/generated-sources/annotations" />
-        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
-        <outputRelativeToContentRoot value="true" />
-        <module name="coho-e2e" />
-      </profile>
-    </annotationProcessing>
-    <bytecodeTargetLevel>
-      <module name="coho-e2e" target="1.5" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
