+++ e2e-coho-tests/src/test/java/pages/WaitConfig.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2014-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ * http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package pages;
+
+public class WaitConfig {
+    public static final int DRAWER_ANIMATION_WAIT = 2000;
+    public static final int VIEWPAGE_BEFORE_WAIT = 1000;
+    public static final int VIEWPAGE_BEFORE_AFTER = 4000;
+}
