+++ e2e-coho-tests/src/test/java/pages/WelcomePage.java
@@ -1,0 +1,22 @@
+package pages;
+
+import io.appium.java_client.AppiumDriver;
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.pagefactory.AndroidFindBy;
+
+public class WelcomePage extends BasePage{
+
+    public WelcomePage(AppiumDriver driver) {
+        super(driver);
+    }
+
+    @AndroidFindBy(xpath = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup[1]")
+    private MobileElement NextButton;
+
+    public IntroductionPage NavigateToIntroductionPage() {
+        NextButton.click();
+        return new IntroductionPage(driver);
+    }
+
+
+}
