+++ pom.xml
@@ -7,9 +7,17 @@
     <groupId>com.verizon.mpersapp.appium.test</groupId>
     <artifactId>Mpersapp-Appium-Test</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <modules>
+        <module>e2e-coho-tests</module>
+    </modules>
+    <packaging>pom</packaging>
     <name>MpersAppAppiumTests</name>
     <url>http://maven.apache.org</url>
+    <properties>
+        <spring.version>5.1.6.RELEASE</spring.version>
+        <spring.ws>3.0.7.RELEASE</spring.ws>
+        <janino.version>2.5.10</janino.version>
+    </properties>
 
     <dependencies>
         <dependency>
@@ -34,6 +42,65 @@
             <artifactId>cucumber-testng</artifactId>
             <version>1.2.4</version>
         </dependency>
+        <!--Spring-->
+        <!--<dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>${spring.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-core</artifactId>
+            <version>${spring.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-test</artifactId>
+            <version>${spring.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.ws</groupId>
+            <artifactId>spring-ws-core</artifactId>
+            <version>${spring.ws}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>commons-logging</groupId>
+                    <artifactId>commons-logging</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-tx</artifactId>
+            <version>${spring.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>janino</groupId>
+            <artifactId>janino</artifactId>
+            <version>${janino.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.slf4j</groupId>
+                    <artifactId>slf4j-log4j12</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>log4j</groupId>
+                    <artifactId>log4j</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>-->
     </dependencies>
 
     <build>
@@ -80,7 +147,7 @@
                             <finalName>zip-with-dependencies</finalName>
                             <appendAssemblyId>false</appendAssemblyId>
                             <descriptors>
-                                <descriptor>src/main/assembly/zip.xml</descriptor>
+                                <descriptor>e2e-coho-tests/src/main/assembly/zip.xml</descriptor>
                             </descriptors>
                         </configuration>
                     </execution>
