--- src/test/java/Pages/GetStartedPage.java
@@ -1,20 +1,0 @@
-package Pages;
-
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.MobileElement;
-import io.appium.java_client.pagefactory.AndroidFindBy;
-
-public class GetStartedPage extends BasePage {
-
-    protected GetStartedPage(AppiumDriver driver) {
-        super(driver);
-    }
-
-    @AndroidFindBy(xpath = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup[5]/android.view.ViewGroup[1]")
-    private MobileElement signInButton;
-
-    public SignInPage NavigateToSignInPage() {
-        signInButton.click();
-        return new SignInPage(driver);
-    }
-}
