--- src/test/java/Pages/IntroductionPage.java
@@ -1,20 +1,0 @@
-package Pages;
-
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.MobileElement;
-import io.appium.java_client.pagefactory.AndroidFindBy;
-
-public class IntroductionPage extends BasePage{
-
-    protected IntroductionPage(AppiumDriver driver) {
-        super(driver);
-    }
-
-    @AndroidFindBy(xpath = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup[2]/android.view.ViewGroup[1]/android.view.ViewGroup")
-    private MobileElement NextButton;
-
-    public GetStartedPage NavigateToGetStartedPage() {
-        NextButton.click();
-        return new GetStartedPage(driver);
-    }
-}
